refactor: Use component instead of container for viewing reviews
